Move FK exception tests to their own class
